SslTest: split `match_domain()` specific tests to dedicated test class
