+++ tests/Ssl/MatchDomainTest.php
@@ -1,0 +1,67 @@
+<?php
+
+namespace WpOrg\Requests\Tests\Ssl;
+
+use WpOrg\Requests\Exception\InvalidArgument;
+use WpOrg\Requests\Ssl;
+use WpOrg\Requests\Tests\Ssl\SslTestCase;
+use WpOrg\Requests\Tests\TypeProviderHelper;
+
+/**
+ * @covers \WpOrg\Requests\Ssl::match_domain
+ */
+final class MatchDomainTest extends SslTestCase {
+
+	/**
+	 * Tests receiving an exception when an invalid input type is passed.
+	 *
+	 * @dataProvider dataInvalidInputType
+	 *
+	 * @param mixed $input Input data.
+	 *
+	 * @return void
+	 */
+	public function testInvalidInputType($input) {
+		$this->expectException(InvalidArgument::class);
+		$this->expectExceptionMessage('Argument #1 ($host) must be of type string|Stringable');
+
+		Ssl::match_domain($input, 'reference');
+	}
+
+	/**
+	 * Data Provider.
+	 *
+	 * @return array
+	 */
+	public function dataInvalidInputType() {
+		return TypeProviderHelper::getAllExcept(TypeProviderHelper::GROUP_STRINGABLE);
+	}
+
+	/**
+	 * Test handling of matching host and DNS names.
+	 *
+	 * @dataProvider dataMatch
+	 *
+	 * @param string $host      Host name to verify.
+	 * @param string $reference DNS name to match against.
+	 *
+	 * @return void
+	 */
+	public function testMatch($host, $reference) {
+		$this->assertTrue(Ssl::match_domain($host, $reference));
+	}
+
+	/**
+	 * Test handling of non-matching host and DNS names.
+	 *
+	 * @dataProvider dataNoMatch
+	 *
+	 * @param string $host      Host name to verify.
+	 * @param string $reference DNS name to match against.
+	 *
+	 * @return void
+	 */
+	public function testNoMatch($host, $reference) {
+		$this->assertFalse(Ssl::match_domain($host, $reference));
+	}
+}
